mp3のダウンロードに対応

+++ src/main.py
@@ -5,6 +5,7 @@ from pathlib import Path
 
 from flask import Flask, Response, jsonify, request
 from moviepy.editor import AudioFileClip, VideoFileClip
+from waitress import serve
 
 # https://github.com/yt-dlp/yt-dlp/blob/master/yt_dlp/YoutubeDL.py
 from yt_dlp import YoutubeDL
@@ -47,12 +48,16 @@ def merge_video(video: str, audio: str, out: str) -> None:
 @app.route("/dl", methods=["GET"])
 async def endpoint() -> tuple[Response, int]:
     try:
-        url = request.args.get("url")
-        fmt = "mp4"
+        query = request.args
+        url = query["url"]
+        fmt = "mp4" if query.get("fmt") is None else query.get("fmt")
+        print("INFO: url =", url, " / fmt =", fmt)
+
         if url is None:
             raise ParameterError
-        ydl = YoutubeDL()
+
         ## ydlインスタンスから情報を収集
+        ydl = YoutubeDL()
         info_dict = get_info(ydl, url)
         [video, audio, title] = get_best_format(info_dict, fmt)
 
@@ -73,20 +78,26 @@ async def endpoint() -> tuple[Response, int]:
             Path(output_name).unlink()
 
         ## 音声/動画共にある場合のみマージ
-        if Path(video_name).exists() and Path(audio_name).exists():
+        if (
+            video_name is not None
+            and audio_name is not None
+            and Path(video_name).exists()
+            and Path(audio_name).exists()
+        ):
             merge_video(
                 video=video_name,
                 audio=audio_name,
                 out=output_name,
             )
-        elif Path(video_name).exists():
+            ## マージ前のメディアを削除
+            for file in video_name, audio_name:
+                if file is not None and Path(file).exists():
+                    Path(file).unlink()
+        ## データが片方しか取得できなかった場合は、リネーム
+        elif video_name is not None and Path(video_name).exists():
             Path(video_name).rename(output_name)
-        elif Path(audio_name).exists():
+        elif audio_name is not None and Path(audio_name).exists():
             Path(audio_name).rename(output_name)
-
-        for file in video_name, audio_name:
-            if Path(file).exists():
-                Path(file).unlink()
 
         ## レスポンス
         result_response = (
@@ -101,7 +112,9 @@ async def endpoint() -> tuple[Response, int]:
             ),
             200,
         )
+        print("INFO: Processes Succeed! ->", output_name)
     except ParameterError:
+        print("Error: Invalid request requested.")
         result_response = (
             jsonify(
                 {
@@ -111,22 +124,14 @@ async def endpoint() -> tuple[Response, int]:
             ),
             400,
         )
-    except BaseException:
-        result_response = (
-            jsonify(
-                {
-                    "error": "Unexpected Failed",
-                    "message": "Failed to process by unexpected reason.",
-                },
-            ),
-            500,
-        )
     return result_response
 
 
 format_set = {
     "mp4": {"video": "mp4", "audio": "m4a"},
     "webm": {"video": "webm", "audio": "webm"},
+    "mp3": {"video": None, "audio": "mp3"},
+    "opus": {"video": None, "audio": "opus"},
 }
 
 
@@ -148,20 +153,34 @@ def get_best_format(info_dict: dict, fmt: str) -> tuple[dict | None, dict | None
     ## ytdlpに受け渡す情報を作成する
     best_video = {
         "format": None,
-        "outtmpl": str(title + " video." + format_set.get(fmt).get("video")),
+        "outtmpl": None,
     }
     best_audio = {
         "format": None,
-        "outtmpl": str(title + " audio." + format_set.get(fmt).get("audio")),
+        "outtmpl": None,
     }
     # 結果の作成
     for dlfmt in info_dict["formats"]:
-        if dlfmt.get("video_ext") == format_set.get(fmt).get("video"):
-            best_video["format"] = dlfmt.get("format_id")
-        if dlfmt.get("audio_ext") == format_set.get(fmt).get("audio"):
-            best_audio["format"] = dlfmt.get("format_id")
+        if (format_set.get(fmt).get("video") is not None) and (
+            dlfmt.get("video_ext") == format_set.get(fmt).get("video")
+        ):
+            best_video["format"] = dlfmt.get("video_ext")
+            best_video["outtmpl"] = str(
+                title + " video." + format_set.get(fmt).get("video"),
+            )
+
+        if (format_set.get(fmt).get("audio") is not None) and (
+            dlfmt.get("audio_ext") == format_set.get(fmt).get("audio")
+        ):
+            best_audio["format"] = dlfmt.get("audio_ext")
+            best_audio["outtmpl"] = str(
+                title + " audio." + format_set.get(fmt).get("audio"),
+            )
     return best_video, best_audio, title
 
 
 if __name__ == "__main__":
-    app.run(host="0.0.0.0", port=5000)
+    print("INFO: Success up ytdlpServer")
+
+    # dev -> app.run(host="0.0.0.0", port=5000)
+    serve(app, host="0.0.0.0", port=5000)
